backend/requirements.txt
boto3
fastapi
PyMySQL
python-dotenv
python-multipart
SQLAlchemy
uvicorn